Move color helpers to own module
