+++ src/helpers/index.js
@@ -1,9 +1,8 @@
 'use strict';
 
-import color from 'chartjs-color';
-
 import * as coreHelpers from './helpers.core';
 import * as canvas from './helpers.canvas';
+import * as color from './helpers.color';
 import * as curve from './helpers.curve';
 import * as dom from './helpers.dom';
 import * as easing from './helpers.easing';
@@ -11,20 +10,10 @@ import * as options from './helpers.options';
 import * as math from './helpers.math';
 import * as rtl from './helpers.rtl';
 
-const colorHelper =
-	function(value) {
-		if (value instanceof CanvasGradient || value instanceof CanvasPattern) {
-			// TODO: figure out what this should be. Previously returned
-			// the default color
-			return value;
-		}
-
-		return color(value);
-	};
-
 export default {
 	...coreHelpers,
 	canvas,
+	color,
 	curve,
 	dom,
 	easing,
@@ -44,11 +33,5 @@ export default {
 	// -- Canvas methods
 	fontString: function(pixelSize, fontStyle, fontFamily) {
 		return fontStyle + ' ' + pixelSize + 'px ' + fontFamily;
-	},
-	color: colorHelper,
-	getHoverColor: function(colorValue) {
-		return (colorValue instanceof CanvasPattern || colorValue instanceof CanvasGradient) ?
-			colorValue :
-			colorHelper(colorValue).saturate(0.5).darken(0.1).rgbString();
 	}
 };
